requirements.txt
python-telegram-bot==20.5
APScheduler==3.10.4
python-dotenv
